ステップ21 引き算を追加。モジュール構成を見直し、演算ごとにモジュールファイルを分割
