+++ src/modules/math/add.rs
@@ -1,0 +1,376 @@
+use crate::modules::functions::*;
+use crate::modules::math::exp::*;
+use crate::modules::math::sub::*;
+use crate::modules::settings::*;
+use crate::modules::variable::*;
+use core::fmt::Debug;
+use ndarray::{Array, IxDyn};
+use std::cell::RefCell;
+use std::ops::Add;
+use std::rc::Rc;
+
+/// 加算関数
+#[derive(Debug, Clone)]
+pub struct AddFunction;
+impl<V: MathOps> Function<V> for AddFunction {
+    // Add (加算) の順伝播
+    fn forward(&self, xs: Vec<Array<V, IxDyn>>) -> Vec<Array<V, IxDyn>> {
+        let result = vec![&xs[0] + &xs[1]];
+        result
+    }
+
+    /// 逆伝播
+    /// y=x0+x1 の微分であるため、dy/dx0=1, dy/dx1=1 である。
+    fn backward(
+        &self,
+        _inputs: Vec<Variable<V>>,
+        gys: Vec<Array<V, IxDyn>>,
+    ) -> Vec<Array<V, IxDyn>> {
+        vec![gys[0].clone(), gys[0].clone()]
+    }
+}
+
+/// 加算関数
+///
+/// Arguments
+/// * x1 (Variable<V>): 加算する変数
+/// * x2 (Variable<V>): 加算する変数
+///
+/// Return
+/// * Rc<RefCell<RawVariable>>: 加算結果
+pub fn add<V: MathOps>(x1: Variable<V>, x2: Variable<V>) -> Variable<V> {
+    let mut add = FunctionExecutor::new(Rc::new(RefCell::new(AddFunction)));
+    // 加算の順伝播
+    add.forward(vec![x1.clone(), x2.clone()])
+        .get(0)
+        .unwrap()
+        .clone()
+}
+
+/// 加算のオーバーロード
+///
+/// Arguments
+/// * self (Variable<V>): 左オペランド
+/// * rhs (Variable<V>): 右オペランド
+///
+/// Returns
+/// * Variable<V>: 加算結果
+impl<V: MathOps> Add<&Variable<V>> for &Variable<V> {
+    type Output = Variable<V>;
+    fn add(self, rhs: &Variable<V>) -> Variable<V> {
+        // 順伝播
+        let mut add = FunctionExecutor::new(Rc::new(RefCell::new(AddFunction)));
+        let result = add
+            .forward(vec![self.clone(), rhs.clone()])
+            .get(0)
+            .unwrap()
+            .clone();
+        result
+    }
+}
+
+/// 加算のオーバーロード (Variable<V> + Array)
+impl<V: MathOps> Add<&Array<V, IxDyn>> for &Variable<V> {
+    type Output = Variable<V>;
+    fn add(self, rhs: &Array<V, IxDyn>) -> Variable<V> {
+        // 順伝播
+        let rhs_val = Variable::new(RawVariable::new(rhs.clone()));
+        self + &rhs_val
+    }
+}
+
+/// 加算のオーバーロード (Array + Variable<V>)
+impl<V: MathOps> Add<&Variable<V>> for &Array<V, IxDyn> {
+    type Output = Variable<V>;
+    fn add(self, rhs: &Variable<V>) -> Variable<V> {
+        // 順伝播
+        let lhs_val = Variable::new(RawVariable::new(self.clone()));
+        &lhs_val + rhs
+    }
+}
+
+/// Variable と様々な数値とのオーバーロード用のマクロ
+macro_rules! impl_variable_add {
+    ($scalar:ty) => {
+        // Variable<V> + $scalar
+        impl<V: MathOps> Add<$scalar> for &Variable<V> {
+            type Output = Variable<V>;
+
+            fn add(self, rhs: $scalar) -> Variable<V> {
+                // 順伝播
+                let rhs_val = Variable::new(RawVariable::new(V::from(rhs).unwrap()));
+                self + &rhs_val
+            }
+        }
+
+        // $scalar + Variable<V>
+        impl<V: MathOps> Add<&Variable<V>> for $scalar {
+            type Output = Variable<V>;
+
+            fn add(self, rhs: &Variable<V>) -> Variable<V> {
+                // 順伝播
+                let lhs_val = Variable::new(RawVariable::new(V::from(self).unwrap()));
+                &lhs_val + rhs
+            }
+        }
+    };
+}
+
+// 複数の数値型に対して加算マクロの一括実装
+impl_variable_add!(i32);
+impl_variable_add!(i64);
+impl_variable_add!(f32);
+impl_variable_add!(f64);
+impl_variable_add!(u32);
+impl_variable_add!(u64);
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use rand::prelude::*;
+
+    /// 加算のテスト
+    #[test]
+    fn test_add() {
+        // 加算値をランダムに生成する。
+        let mut rng = rand::rng();
+        let rand_x1 = rng.random::<f64>();
+        let rand_x2 = rng.random::<f64>();
+        let x1 = Variable::new(RawVariable::new(rand_x1));
+        let x2 = Variable::new(RawVariable::new(rand_x2));
+
+        // 加算した結果の期待値を計算する。
+        let expected_output_data = Array::from_elem(IxDyn(&[]), rand_x1 + rand_x2);
+
+        // 順伝播、逆伝播を実行する。
+        // 逆伝播のため、順伝播の関数の実行結果を取得し、逆伝播を実行する。
+        let result = add(x1, x2);
+
+        // 足し算の結果
+        assert_eq!(expected_output_data, result.borrow().get_data());
+    }
+
+    /// オーバーロードのテスト
+    #[test]
+    fn test_add_mul_overload() {
+        // 逆伝播を実行する。微分値を保持する。
+        Setting::set_retain_grad_enabled();
+
+        // バックプロパゲーションを行う。
+        Setting::set_backprop_enabled();
+
+        // 変数を用意する。
+        let mut raw_a = RawVariable::new(3.0f32);
+        raw_a.set_name("val_a".to_string());
+        let a = Variable::new(raw_a);
+
+        let mut raw_b = RawVariable::new(2.0f32);
+        raw_b.set_name("val_b".to_string());
+        let b = Variable::new(raw_b);
+        let mut raw_c = RawVariable::new(1.0f32);
+        raw_c.set_name("val_c".to_string());
+        let c = Variable::new(raw_c);
+
+        // 計算する。a * b + c
+        let result = &(&a * &b) + &c;
+
+        let expected = RawVariable::new(7.0f32);
+
+        // 逆伝播を実行する。
+        result.backward();
+
+        println!(
+            "result grad: {:?}, a grad: {:?}, b grad: {:?}, c grad: {:?}",
+            &result.borrow().get_grad(),
+            // &a.borrow().get_grad(),
+            &a.borrow().get_grad(),
+            &b.borrow().get_grad(),
+            &c.borrow().get_grad(),
+        );
+
+        assert_eq!(expected.get_data(), result.borrow().get_data());
+        assert_eq!(
+            Array::from_elem(IxDyn(&[]), 1.0),
+            result.borrow().get_grad().expect("No grad exist.")
+        );
+        assert_eq!(
+            Array::from_elem(IxDyn(&[]), 2.0),
+            a.borrow().get_grad().expect("No grad exist.")
+        );
+        assert_eq!(
+            Array::from_elem(IxDyn(&[]), 3.0),
+            b.borrow().get_grad().expect("No grad exist.")
+        );
+    }
+
+    /// オーバーロードのテスト (Array との計算)
+    #[test]
+    fn test_add_mul_with_array_overload() {
+        // 逆伝播を実行する。微分値を保持する。
+        Setting::set_retain_grad_enabled();
+
+        // バックプロパゲーションを行う。
+        Setting::set_backprop_enabled();
+
+        // 変数を用意する。
+        let mut raw_a = RawVariable::new(3i32);
+        raw_a.set_name("val_a".to_string());
+        let a = Variable::new(raw_a);
+
+        // let mut raw_b = RawVariable::new(2i32);
+        // raw_b.set_name("val_b".to_string());
+        // let b = Variable::new(raw_b);
+
+        // b は Array とする。
+        let b = Array::from_elem(IxDyn(&[]), 2i32);
+
+        let mut raw_c = RawVariable::new(1i32);
+        raw_c.set_name("val_c".to_string());
+        let c = Variable::new(raw_c);
+
+        // 計算する。a * b + c
+        let result = &(&a * &b) + &c;
+
+        let expected = RawVariable::new(7i32);
+
+        // 逆伝播を実行する。
+        result.backward();
+
+        println!(
+            "result grad: {:?}, a grad: {:?}, c grad: {:?}",
+            &result.borrow().get_grad(),
+            // &a.borrow().get_grad(),
+            &a.borrow().get_grad(),
+            // &b.borrow().get_grad(),
+            &c.borrow().get_grad(),
+        );
+
+        assert_eq!(expected.get_data(), result.borrow().get_data());
+        assert_eq!(
+            Array::from_elem(IxDyn(&[]), 1),
+            result.borrow().get_grad().expect("No grad exist.")
+        );
+        assert_eq!(
+            Array::from_elem(IxDyn(&[]), 2),
+            a.borrow().get_grad().expect("No grad exist.")
+        );
+        // assert_eq!(
+        //     Array::from_elem(IxDyn(&[]), 3),
+        //     b.borrow().get_grad().expect("No grad exist.")
+        // );
+    }
+
+    /// 乗算オーバーロードのテスト
+    /// 様々な型、および、左右オペランドを入れ替えたテスト
+    #[test]
+    fn test_mul_overload_macro() {
+        let overload_val_i32 = Variable::new(RawVariable::new(2i32));
+        let overload_val_f32 = Variable::new(RawVariable::new(2.0f32));
+        let overload_val_f64 = Variable::new(RawVariable::new(2.0f64));
+        let overload_val_u32 = Variable::new(RawVariable::new(2u32));
+        let overload_array_f32 = Array::from_elem(IxDyn(&[]), 2.0f32);
+
+        let result_val_i32_mul_val_i32 = &overload_val_i32 * &overload_val_i32;
+        let result_val_u32_mul_scalar_u32 = &overload_val_u32 * 10u32;
+        let result_scalar_f64_mul_val_f64 = 10.0f64 * &overload_val_f64;
+        let result_val_f32_mul_array_f32 = &overload_val_f32 * &overload_array_f32;
+        let result_array_f32_mul_val_f32 = &overload_array_f32 * &overload_val_f32;
+
+        assert_eq!(
+            RawVariable::new(4i32).get_data(),
+            result_val_i32_mul_val_i32.borrow().get_data()
+        );
+
+        assert_eq!(
+            RawVariable::new(20u32).get_data(),
+            result_val_u32_mul_scalar_u32.borrow().get_data()
+        );
+
+        assert_eq!(
+            RawVariable::new(20.0f64).get_data(),
+            result_scalar_f64_mul_val_f64.borrow().get_data()
+        );
+
+        assert_eq!(
+            RawVariable::new(4.0f32).get_data(),
+            result_val_f32_mul_array_f32.borrow().get_data()
+        );
+
+        assert_eq!(
+            RawVariable::new(4.0f32).get_data(),
+            result_array_f32_mul_val_f32.borrow().get_data()
+        );
+    }
+
+    /// 加算オーバーロードのテスト
+    /// 様々な型、および、左右オペランドを入れ替えたテスト
+    #[test]
+    fn test_add_overload_macro() {
+        let overload_val_i64 = Variable::new(RawVariable::new(2i64));
+        let overload_val_f32 = Variable::new(RawVariable::new(2.0f32));
+        let overload_val_f64 = Variable::new(RawVariable::new(2.0f64));
+        let overload_val_u64 = Variable::new(RawVariable::new(2u64));
+        let overload_array_f32 = Array::from_elem(IxDyn(&[]), 2.0f32);
+
+        let result_val_i64_add_val_i64 = &overload_val_i64 + &overload_val_i64;
+        let result_val_u64_add_scalar_u64 = &overload_val_u64 + 10u64;
+        let result_scalar_f64_add_val_f64 = 10.0f64 + &overload_val_f64;
+        let result_val_f32_add_array_f32 = &overload_val_f32 + &overload_array_f32;
+        let result_array_f32_add_val_f32 = &overload_array_f32 + &overload_val_f32;
+
+        assert_eq!(
+            RawVariable::new(4i64).get_data(),
+            result_val_i64_add_val_i64.borrow().get_data()
+        );
+
+        assert_eq!(
+            RawVariable::new(12u64).get_data(),
+            result_val_u64_add_scalar_u64.borrow().get_data()
+        );
+
+        assert_eq!(
+            RawVariable::new(12.0f64).get_data(),
+            result_scalar_f64_add_val_f64.borrow().get_data()
+        );
+
+        assert_eq!(
+            RawVariable::new(4.0f32).get_data(),
+            result_val_f32_add_array_f32.borrow().get_data()
+        );
+
+        assert_eq!(
+            RawVariable::new(4.0f32).get_data(),
+            result_array_f32_add_val_f32.borrow().get_data()
+        );
+    }
+
+    /// 負数 Neg に関するテスト
+    #[test]
+    fn test_neg_overload() {
+        // 逆伝播を実行する。微分値を保持する。
+        Setting::set_retain_grad_enabled();
+
+        // バックプロパゲーションを行う。
+        Setting::set_backprop_enabled();
+
+        let pos_val_i32_1 = Variable::new(RawVariable::new(2i32));
+        let pos_val_i32_2 = Variable::new(RawVariable::new(3i32));
+        let pos_val_i32_3 = Variable::new(RawVariable::new(4i32));
+        let neg_val_i32 = &(&pos_val_i32_1 + &-pos_val_i32_2.clone()) + &-pos_val_i32_3.clone();
+
+        assert_eq!(
+            RawVariable::new(-5).get_data(),
+            &neg_val_i32.borrow().get_data()
+        );
+
+        let pos_val_f64_1 = Variable::new(RawVariable::new(2f64));
+        let pos_val_f64_2 = Variable::new(RawVariable::new(3f64));
+        let pos_val_f64_3 = Variable::new(RawVariable::new(4f64));
+        let neg_val_f64 = &(&pos_val_f64_1 + &-pos_val_f64_2) + &-pos_val_f64_3;
+
+        assert_eq!(
+            RawVariable::new(-5f64).get_data(),
+            &neg_val_f64.borrow().get_data()
+        );
+    }
+}
