+++ src/modules/math/exp.rs
@@ -1,0 +1,70 @@
+use crate::modules::functions::*;
+use crate::modules::math::add::*;
+use crate::modules::settings::*;
+use crate::modules::variable::*;
+use core::fmt::Debug;
+use ndarray::{Array, IxDyn};
+use std::cell::RefCell;
+use std::rc::Rc;
+
+/// Exp 関数
+#[derive(Debug, Clone)]
+pub struct ExpFunction;
+impl<V: MathOps> Function<V> for ExpFunction {
+    // Exp (y=e^x) の順伝播
+    fn forward(&self, xs: Vec<Array<V, IxDyn>>) -> Vec<Array<V, IxDyn>> {
+        let e = std::f64::consts::E;
+        let result = vec![xs[0].mapv(|x| V::from(e.powf(x.to_f64().unwrap())).unwrap())];
+
+        result
+    }
+
+    /// 逆伝播
+    /// dy/dx=e^x である。
+    fn backward(
+        &self,
+        inputs: Vec<Variable<V>>,
+        gys: Vec<Array<V, IxDyn>>,
+    ) -> Vec<Array<V, IxDyn>> {
+        let e = std::f64::consts::E;
+        let x = inputs[0].borrow().get_data();
+        let gys_val = gys[0].clone();
+        let x_exp = vec![x.mapv(|x| V::from(e.powf(x.to_f64().unwrap())).unwrap())];
+        let gxs = x_exp.iter().map(|x_exp| x_exp * &gys_val).collect();
+        gxs
+    }
+}
+
+/// Exp 関数
+///
+/// Arguments
+/// * input (Rc<RefCell<RawVariable>>): 入力値
+///
+/// Return
+/// * Rc<RefCell<RawVariable>>: 結果
+pub fn exp<V: MathOps>(input: Variable<V>) -> Variable<V> {
+    let mut exp = FunctionExecutor::new(Rc::new(RefCell::new(ExpFunction)));
+    // EXP の順伝播
+    exp.forward(vec![input.clone()]).get(0).unwrap().clone()
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use rand::prelude::*;
+
+    /// Exp 関数のテスト。
+    #[test]
+    fn test_exp() {
+        let x = Variable::new(RawVariable::new(2.0));
+
+        let e = std::f64::consts::E;
+        let expected_output_data = Array::from_elem(IxDyn(&[]), e.powf(2.0));
+
+        // 順伝播、逆伝播を実行する。
+        let result = exp(x);
+
+        // exp 結果
+        assert_eq!(expected_output_data, result.borrow().get_data());
+    }
+}
